Split autoimpl tests
